NavMesh Hiding Prototyping
Prototyped useful hiding mechanic for NavMeshAgent based on the following video:
https://www.youtube.com/watch?v=t9e2XBQY4Og&ab_channel=LlamAcademy

Current code implementation is near identical and was meant for learning and understanding the process. Changes made primarily include renaming variables and adding in comments. The next intended iteration will likely be a recreation of this code in a better adapted way to the project.